+++ CHANGELOG.md
@@ -24,6 +24,10 @@ tech changes will usually be stripped from release notes for the public
     - Add effects with infinite timespan
     - Fix names in initiative list not being reactive
 
+### Fixed
+
+-   Images in the Token Direction indicator were overflowing
+
 ## [2025.3]
 
 ### Added